refactor: reuse grading and completion logic
